requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
passlib==1.7.4
pypdf==4.3.1
python-docx==1.1.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
google-generativeai==0.8.3
transformers==4.45.2
torch==2.4.1
sentencepiece==0.2.0
faiss-cpu==1.8.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
